--- src/main/java/org/joelson/turf/turfgame/apiv5util/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Utility classes to manipulate data from http://api.turfgame.com/unstable .
- */
-package org.joelson.turf.turfgame.apiv5util;
